Setup website structure
Big first commit. Website is layed-out with two columns; list of works
on the left, the work's details on the right. Swapping between projects
is implemented and so is a basic image gallery, bar a few minor bugs.

Most of the functionality here is pulled from my experiments in the
'index' repo with the goal of compiling them into a lightweight, single
page site.